Revert "changed all getCurrentRating() to getGlickoRating(); also added the Impl to my old classes"
This reverts commit c53b43c8b52b796e32dae00565abbfbf736f2cb5.
