+++ backend/src/main/java/com/project/G1_T3/leaderboard/model/LeaderboardPlayerProfile.java
@@ -32,7 +32,7 @@ public class LeaderboardPlayerProfile {
         profileId = player.getProfileId();
         firstName = player.getFirstName();
         lastName = player.getLastName();
-        ELO = player.getGlickoRating();
+        ELO = player.getCurrentRating();
         position = p;
     }
