+++ backend/src/main/java/com/project/G1_T3/stage/service/StageServiceImpl.java
@@ -92,7 +92,7 @@ public class StageServiceImpl implements StageService {
     
         // Sort players by rating
         List<PlayerProfile> sortedPlayers = new ArrayList<>(players);
-        Collections.sort(sortedPlayers, (a, b) -> Double.compare(b.getGlickoRating(), a.getGlickoRating()));
+        Collections.sort(sortedPlayers, (a, b) -> Double.compare(b.getCurrentRating(), a.getCurrentRating()));
     
         // Create the first round with sorted players
         try {
